Allow multiple convertions in the same text node
Also solves a problem where for example: (12 pounds) would screw up
since the "fullMatch" would be '12 pounds)', when using with replace
(which is used before this commit) it would throw an error since it
takes it as an invalid regex (closes parenthesis but doesn't open one).

Should also be a bit faster now when not using regex to replace/insert
the new value.